// models/sales_calendar.model.ts
export type SalesCalendarState = {
  data: {
    sals_cale_id: string;
    sals_cale_date: string;
    sals_cale_fullname: string;
    sals_cale_company_name: string;
    sals_cale_tel: string;
    sals_cale_email: string;
    sals_cale_doc_type: string;
    sals_cale_printing_type: string;
    sals_cale_amount: string;
    sals_cale_quotation_request: string;
    sals_cale_finished_size: string;
    sals_cale_page: string;
    sals_cale_paper: string;
    sals_cale_printing: string;
    sals_cale_stand: string;
    sals_cale_binding: string;
    slas_cale_o_ring_color: String;
    sals_cale_printing_volume: string;
  };
};